+++ elide-datastore/elide-datastore-aggregation/src/test/java/com/yahoo/elide/datastores/aggregation/framework/AggregationDataStoreTestHarness.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.datastores.aggregation.framework;
 
-import com.yahoo.elide.core.datastore.DataStore;
 import com.yahoo.elide.core.datastore.test.DataStoreTestHarness;
 import com.yahoo.elide.core.utils.ClassScanner;
 import com.yahoo.elide.core.utils.DefaultClassScanner;
@@ -20,10 +19,10 @@ import com.yahoo.elide.datastores.aggregation.queryengines.sql.dialects.SQLDiale
 import com.yahoo.elide.datastores.aggregation.queryengines.sql.query.AggregateBeforeJoinOptimizer;
 import com.yahoo.elide.datastores.jpa.JpaDataStore;
 import com.yahoo.elide.datastores.jpa.transaction.NonJtaTransaction;
-import com.yahoo.elide.datastores.multiplex.MultiplexManager;
 import com.yahoo.elide.modelconfig.validator.DynamicConfigValidator;
 import org.hibernate.Session;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -35,7 +34,8 @@ import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 @AllArgsConstructor
-public class AggregationDataStoreTestHarness implements DataStoreTestHarness {
+@Getter
+public abstract class AggregationDataStoreTestHarness implements DataStoreTestHarness {
     private EntityManagerFactory entityManagerFactory;
     private ConnectionDetails defaultConnectionDetails;
     private Map<String, ConnectionDetails> connectionDetailsMap;
@@ -87,18 +87,6 @@ public class AggregationDataStoreTestHarness implements DataStoreTestHarness {
                         new DefaultQueryPlanMerger(metaDataStore),
                         new DefaultQueryValidator(metaDataStore.getMetadataDictionary())))
                 .queryLogger(new Slf4jQueryLogger());
-    }
-
-    @Override
-    public DataStore getDataStore() {
-
-        MetaDataStore metaDataStore = createMetaDataStore();
-
-        AggregationDataStore.AggregationDataStoreBuilder aggregationDataStoreBuilder = createAggregationDataStoreBuilder(metaDataStore);
-
-        DataStore jpaStore = createJPADataStore();
-
-        return new MultiplexManager(jpaStore, metaDataStore, aggregationDataStoreBuilder.build());
     }
 
     @Override
